requirements.txt
sentence-transformers
qdrant-client
python-dotenv